Deleted template directory and moved popup templates up to parent directory
